+++ packages/client/src/observable/ObservableClient.ts
@@ -148,6 +148,7 @@ export interface ObserveObjectsCallbackArgs<
   fetchMore: () => Promise<void>;
   hasMore: boolean;
   status: Status;
+  totalCount?: string;
 }
 
 export interface ObserveObjectSetArgs<
@@ -166,6 +167,7 @@ export interface ObserveObjectSetArgs<
   hasMore: boolean;
   status: Status;
   objectSet: ObjectSet<T, RDPs>;
+  totalCount?: string;
 }
 
 export interface ObserveAggregationOptions<
@@ -227,6 +229,7 @@ export interface ObserveLinkCallbackArgs<
   fetchMore: () => Promise<void>;
   hasMore: boolean;
   status: Status;
+  totalCount?: string;
 }
 
 /**