+++ packages/client/src/observable/internal/objectset/ObjectSetQuery.ts
@@ -204,7 +204,12 @@ export class ObjectSetQuery extends BaseListQuery<
     this.logger?.error("error", error);
     this.store.subjects.get(this.cacheKey).error(error);
 
-    return this.writeToStore({ data: [] }, "error", batch);
+    const existingTotalCount = batch.read(this.cacheKey)?.value?.totalCount;
+    return this.writeToStore(
+      { data: [], totalCount: existingTotalCount },
+      "error",
+      batch,
+    );
   }
 
   registerStreamUpdates(sub: Subscription): void {
@@ -232,6 +237,7 @@ export class ObjectSetQuery extends BaseListQuery<
       isOptimistic: boolean;
       status: Status;
       lastUpdated: number;
+      totalCount?: string;
     },
   ): ObjectSetPayload {
     return {
@@ -242,6 +248,7 @@ export class ObjectSetQuery extends BaseListQuery<
       status: params.status,
       lastUpdated: params.lastUpdated,
       objectSet: this.#composedObjectSet,
+      totalCount: params.totalCount,
     };
   }
 }