+++ packages/create-widget.template.react.v2/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @osdk/create-widget.template.react.v2
 
+## 3.4.0-beta.14
+
+## 3.4.0-beta.13
+
 ## 3.4.0-beta.12
 
 ### Minor Changes