+++ packages/create-widget/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @osdk/create-widget
 
+## 3.4.0-beta.14
+
+## 3.4.0-beta.13
+
 ## 3.4.0-beta.12
 
 ## 3.4.0-beta.11